display_manager/requirements/requirements.txt
luma.core==2.4.1
luma.oled==3.13.0
Pillow==10.0.1
RPi.GPIO
pika
